refactor(prerender): extract handler.js into focused modules
Move functions from handler.js into dedicated modules by concern:
- utils/opportunity-utils.js: suggestion/opportunity management (9 functions)
- utils/url-selector.js: URL candidate assembly (5 functions)
- utils/scrape-utils.js: S3 content access and scrape job helpers (7 functions)
- utils/citability.js: page citability record writes (1 function)
- ai-only-handler.js: Slack-triggered reconciliation mode (1 function)
- sendPrerenderGuidanceRequestToMystique moved into guidance-handler.js

handler.js is now a thin orchestrator of exported step functions.
