+++ src/prerender/ai-only-handler.js
@@ -1,0 +1,139 @@
+/*
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import { Audit } from '@adobe/spacecat-shared-data-access';
+import { fetchLatestScrapeJobId } from './utils/scrape-utils.js';
+import { sendPrerenderGuidanceRequestToMystique } from './guidance-handler.js';
+import { MODE_AI_ONLY } from './utils/constants.js';
+
+const LOG_PREFIX = 'Prerender -';
+const AUDIT_TYPE = Audit.AUDIT_TYPES.PRERENDER;
+
+/**
+ * Handles AI-summary-only mode: sends existing suggestions to Mystique without running audit.
+ * Called early in step 1 to bypass import/scraping/processing steps.
+ * @param {Object} context - Audit context
+ * @returns {Promise<Object>} - Result indicating success/failure
+ */
+export async function handleAiOnlyMode(context) {
+  const {
+    site, log, dataAccess, data,
+  } = context;
+  const { Opportunity } = dataAccess;
+  const siteId = site.getId();
+  const baseUrl = site.getBaseURL();
+
+  // Parse optional params from data field (opportunityId, scrapeJobId)
+  let opportunityId = null;
+  let scrapeJobId = null;
+  if (data) {
+    try {
+      const parsedData = typeof data === 'string' ? JSON.parse(data) : data;
+      opportunityId = parsedData.opportunityId;
+      scrapeJobId = parsedData.scrapeJobId;
+    } catch (e) {
+      // Ignore parse errors - graceful degradation for malformed JSON
+    }
+  }
+
+  log.info(`${LOG_PREFIX} ai-only: Processing AI summary request for baseUrl=${baseUrl}, siteId=${siteId}, opportunityId=${opportunityId || 'latest'}`);
+
+  // Fetch scrapeJobId from status.json if not provided
+  if (!scrapeJobId) {
+    log.info(`${LOG_PREFIX} ai-only: scrapeJobId not provided, fetching from status.json for baseUrl=${baseUrl}, siteId=${siteId}`);
+    scrapeJobId = await fetchLatestScrapeJobId(siteId, context);
+
+    if (!scrapeJobId) {
+      const error = 'scrapeJobId not found. Either provide it in data or ensure a prerender audit has run recently.';
+      log.error(`${LOG_PREFIX} ai-only: ${error} baseUrl=${baseUrl}, siteId=${siteId}`);
+      return {
+        error,
+        status: 'failed',
+        fullAuditRef: `${MODE_AI_ONLY}/failed-${siteId}`,
+        auditResult: { error },
+      };
+    }
+  }
+
+  // Find the opportunity
+  let opportunity;
+  if (opportunityId) {
+    opportunity = await Opportunity.findById(opportunityId);
+    if (!opportunity) {
+      const error = `Opportunity not found: ${opportunityId}`;
+      log.error(`${LOG_PREFIX} ai-only: ${error} baseUrl=${baseUrl}, siteId=${siteId}`);
+      return {
+        error,
+        status: 'failed',
+        fullAuditRef: `${MODE_AI_ONLY}/failed-${siteId}`,
+        auditResult: { error },
+      };
+    }
+  } else {
+    // Find latest NEW prerender opportunity for this site
+    const opportunities = await Opportunity.allBySiteIdAndStatus(siteId, 'NEW');
+    opportunity = opportunities.find((o) => o.getType() === AUDIT_TYPE);
+
+    if (!opportunity) {
+      const error = `No NEW prerender opportunity found for site: ${siteId}`;
+      log.error(`${LOG_PREFIX} ai-only: ${error} baseUrl=${baseUrl}, siteId=${siteId}`);
+      return {
+        error,
+        status: 'failed',
+        fullAuditRef: `${MODE_AI_ONLY}/failed-${siteId}`,
+        auditResult: { error },
+      };
+    }
+
+    log.info(`${LOG_PREFIX} ai-only: Found latest NEW opportunity: ${opportunity.getId()} for baseUrl=${baseUrl}, siteId=${siteId}`);
+  }
+
+  // Verify opportunity belongs to the site
+  if (opportunity.getSiteId() !== siteId) {
+    const error = `Opportunity ${opportunity.getId()} does not belong to site ${siteId}`;
+    log.error(`${LOG_PREFIX} ai-only: ${error} baseUrl=${baseUrl}, siteId=${siteId}`);
+    return {
+      error,
+      status: 'failed',
+      fullAuditRef: `${MODE_AI_ONLY}/failed-${siteId}`,
+      auditResult: { error },
+    };
+  }
+
+  // Send to Mystique using the existing function
+  const auditData = {
+    siteId,
+    // Fallback to custom audit ID for ai-only mode (for old opportunities without auditId)
+    auditId: opportunity.getAuditId() || `prerender-ai-only-${siteId}`,
+    scrapeJobId,
+  };
+
+  const suggestionCount = await sendPrerenderGuidanceRequestToMystique(
+    site.getBaseURL(),
+    auditData,
+    opportunity,
+    context,
+  );
+
+  log.info(`${LOG_PREFIX} ai-only: Successfully queued AI summary request for ${suggestionCount} suggestion(s). baseUrl=${baseUrl}, siteId=${siteId}, opportunityId=${opportunity.getId()}`);
+
+  return {
+    status: 'complete',
+    mode: MODE_AI_ONLY,
+    opportunityId: opportunity.getId(),
+    fullAuditRef: `${MODE_AI_ONLY}/${opportunity.getId()}`,
+    auditResult: {
+      message: `AI summary generation queued successfully for ${suggestionCount} suggestion(s)`,
+      suggestionCount,
+    },
+  };
+}
